Created a getTemplatePath to allow me to have partial views displayed in the static files and live reload files.
Created an orientation class.

Created the phone service and added some implementation to the radio service.

Corrected an issue where I couldn't capture the touchmove event without killing the default touchstart event.

Added some of the hooks that will be needed in order for cordova to call into the angular application.