Fixed the yaw indicator for the copter.
It seems that the initial load results in the copter reporting Yaw as NaN to the directive. Because I was checking for the existence of scope.angle and throwing an exception if it didn't exist the rest of the directive was never ran. I now check for attribute.angle instead and just accept that initially it will be NaN.